requirements.txt
aniso8601==9.0.1
click==8.1.3
flake8==6.0.0
Flask==2.2.2
graphene==3.2.1
graphql-core==3.2.3
graphql-relay==3.2.0
graphql-server==3.0.0b5
importlib-metadata==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
psycopg2==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
python-dotenv==0.21.0
typing_extensions==4.4.0
Werkzeug==2.2.2
zipp==3.11.0
